Rename file of useDate to getDateinfo
