+++ src/utils/DateUtils/index.ts
@@ -1,1 +1,1 @@
-export { useDate } from './useDate';
+export { useDate } from './getDateInfo';
